dom/tests/mochitest/w3c/html/tests/submission/Mozilla/test_window-onerror-runtime-error.html

<!doctype html>
<html>
 <head>
  <title>window.onerror: runtime scripterrors</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <!--

    http://www.w3.org/TR/html5/webappapis.html#runtime-script-errors
    says what to do for uncaught runtime script errors, and just below
    describes what to do when onerror is a Function.

    -->
 </head>
 <body>

  <div id="log"></div>
  <script>
    var error_count = 0;
    window.onerror = function(msg, url, lineno) {
      ++error_count;
      test(function() {assert_equals(url, window.location.href)},
           "correct url passed to window.onerror");
      test(function() {assert_equals(lineno, 35)},
           "correct line number passed to window.onerror");
    };
  </script>
  <script>
  try {
    // This error is caught, so it should NOT trigger onerror.
    window.nonexistentproperty.oops();
  } catch (ex) {
  }
  // This error is NOT caught, so it should trigger onerror.
  window.nonexistentproperty.oops();
  </script>
  <script>
  test(function() {assert_equals(error_count, 1)},
       "correct number of calls to window.onerror");
  </script>
 </body>
</html>
